+++ resource.go
@@ -21,6 +21,7 @@ func Search(endpoint string, search string) (result structs.Resource,
 	err error) {
 	err = do(fmt.Sprintf("%s?offset=0&limit=9999", endpoint), &result)
 	result.Results = parseSearch(result.Results, search)
+	result.Count = len(result.Results)
 	return
 }
 
@@ -37,9 +38,21 @@ func parseParams(params []int) (offset int, limit int) {
 
 func parseSearch(results []structs.Result, search string) []structs.Result {
 	var x int
+	var substr string
+
 	for _, result := range results {
-		if !strings.Contains(result.Name, search) {
-			continue
+		if string(search[0]) == "^" {
+			substr = string(search[1:])
+			if len(substr) > len(result.Name) {
+				continue
+			}
+			if string(result.Name[0:len(substr)]) != string(substr) {
+				continue
+			}
+		} else {
+			if !strings.Contains(result.Name, search) {
+				continue
+			}
 		}
 		results[x] = result
 		x++