+++ tests/resource_test.go
@@ -51,3 +51,29 @@ func TestSearch(t *testing.T) {
 	assert.Equal(t, "venusaur", result.Results[2].Name,
 		"Expect to receive Venusaur.")
 }
+
+func TestSearchFail(t *testing.T) {
+	result, _ := pokeapi.Search("pokemon", "asdf")
+	assert.Equal(t, 0, len(result.Results),
+		"Expect to receive zero results.")
+}
+
+func TestSearchStartsWith(t *testing.T) {
+	result, _ := pokeapi.Search("pokemon", "^a")
+	assert.Equal(t, 44, len(result.Results),
+		"Expect to receive four results.")
+	assert.Equal(t, "arbok", result.Results[0].Name,
+		"Expect to receive Arbok.")
+
+	result, _ = pokeapi.Search("pokemon", "^bla")
+	assert.Equal(t, 5, len(result.Results),
+		"Expect to receive four results.")
+	assert.Equal(t, "blastoise", result.Results[0].Name,
+		"Expect to receive Blastoise.")
+}
+
+func TestSearchStartsWithFail(t *testing.T) {
+	result, _ := pokeapi.Search("pokemon", "^zzz")
+	assert.Equal(t, 0, len(result.Results),
+		"Expect to receive zero results.")
+}